+++ src/Delegate.Daxif/HelperModules/TypeProviderHelper.fs
@@ -1,12 +1,14 @@
 ﻿namespace DG.Daxif.HelperModules
 
 open System
+open DG.Daxif
 open DG.Daxif.HelperModules.Common
 
 open Microsoft.Xrm.Sdk.Metadata
 open Microsoft.Xrm.Sdk.Client
 
 open ProviderImplementation.ProvidedTypes
+open Microsoft.FSharp.Quotations
 
 module internal TypeProviderHelper =
 
@@ -54,25 +56,39 @@ module internal TypeProviderHelper =
   let createEntityMetadataProps (data:XrmData) (entity:string) () =
     let metadata = data.entityAttributes.Value.[entity].Value
 
-    // Add "static" LogicalName and SchemaName
+    // Add "static" LogicalName and SchemaName as well as Primary Attributes
     ([ ProvidedProperty(
-        ".LogicalName",
+        "(LogicalName)",
         typeof<string>,
         IsStatic = true,
         GetterCode = fun args ->
           let logicalName = metadata.LogicalName 
           <@@ logicalName @@>);
        ProvidedProperty(
-        ".SchemaName",
+        "(SchemaName)",
         typeof<string>,
         IsStatic = true,
         GetterCode = fun args ->
           let schemaName = metadata.SchemaName
-          <@@ schemaName @@>); ]) @
+          <@@ schemaName @@>);
+       ProvidedProperty(
+        "(PrimaryIdAttribute)",
+        typeof<string>,
+        IsStatic = true,
+        GetterCode = fun args ->
+          let primaryIdAttribute = metadata.PrimaryIdAttribute
+          <@@ primaryIdAttribute @@>);
+       ProvidedProperty(
+        "(PrimaryNameAttribute)",
+        typeof<string>,
+        IsStatic = true,
+        GetterCode = fun args ->
+          let primaryNameAttribute = metadata.PrimaryNameAttribute
+          <@@ primaryNameAttribute @@>); ]) @
 
     // Add static array of all attributes
     ([ ProvidedProperty(
-        ".All Attributes",
+        "(All Attributes)",
         typeof<string array>,
         IsStatic=true,
         GetterCode = fun args -> 
@@ -88,8 +104,8 @@ module internal TypeProviderHelper =
           <@@ name @@>))) @
 
     // Add relationship props
-    ([  metadata.OneToManyRelationships |> Array.map box
-        metadata.ManyToOneRelationships |> Array.map box
+    ([  metadata.OneToManyRelationships  |> Array.map box
+        metadata.ManyToOneRelationships  |> Array.map box
         metadata.ManyToManyRelationships |> Array.map box
      ] 
      |> Array.concat |> List.ofArray 
@@ -103,6 +119,22 @@ module internal TypeProviderHelper =
 
 
   let getEntityRecords (data:XrmData) (em:EntityMetadata) () =
+
+    // Add static array of all entities (light)
+    let allRecords () =
+      ProvidedProperty(
+        "(All Records)",
+        typeof<Guid array>,
+        IsStatic=true,
+        GetterCode = fun args -> 
+        let all =
+          data.records.Value.[em.LogicalName].Value
+//          |> Seq.map(fun e -> 
+//            e.Attributes.[em.PrimaryNameAttribute].ToString(),e.Id.ToString())
+//          |> Seq.map(fun (name,id) -> <@@ Tuple(item1 = (name,Guid(id))) @@>)
+          |> Seq.map(fun e -> let id = e.Id.ToString() in <@@ Guid(id) @@>)
+          |> Seq.toList
+        Expr.NewArray(typedefof<Guid>,all))
 
     let records = 
       data.records.Value.[em.LogicalName].Value
@@ -120,7 +152,7 @@ module internal TypeProviderHelper =
       |> List.ofSeq
 
     match records.IsEmpty with
-    | false -> records
+    | false -> allRecords () :: records
     | true  -> 
       [ProvidedProperty("No records available", typeof<obj>, IsStatic=true,
         GetterCode = fun args -> <@@ null @@>)]
@@ -146,10 +178,10 @@ module internal TypeProviderHelper =
     let data = XrmData proxy
 
     let version =
-      ProvidedProperty("Version", typeof<string>, IsStatic=true,
+      ProvidedProperty("Version", typeof<Tuple<string,CrmReleases>>, IsStatic=true,
         GetterCode = fun args -> 
-          let version = data.version.Value
-          <@@ version @@>)
+          let version,release = data.version.Value
+          <@@ (version,release) @@>)
     version.AddXmlDocDelayed(
       fun () -> "Returns the version number of the CRM system.")
 
@@ -172,7 +204,7 @@ module internal TypeProviderHelper =
       fun () -> 
         // Add static array of all entities
         ProvidedProperty(
-            ".All Entities",
+            "(All Entities)",
             typeof<string array>,
             IsStatic=true,
             GetterCode = fun args ->